LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.3_The_for_Statement/S12.6.3_A11_T3-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Parse error
S12.6.3_A11_T3

PASS Expected parsing failure

TEST COMPLETE
